+++ client/src/components/CheckoutForm.jsx
@@ -1,6 +1,8 @@
 import React from 'react'
-
+import { patchOrder } from '../api/orders';
 import { useEffect, useState } from "react";
+import useCart from '../hooks/useCart';
+import useAuth from '../hooks/useAuth';
 import {
   PaymentElement,
   LinkAuthenticationElement,
@@ -17,7 +19,10 @@ export default function CheckoutForm() {
   const [email, setEmail] = useState('');
   const [message, setMessage] = useState(null);
   const [isLoading, setIsLoading] = useState(false);
-
+  const { cart, setCart } = useCart();
+  const {user,setUser} = useAuth();
+  const [validationError,setValidationError]=useState(false);
+  console.log("cart in checkout",cart)
   useEffect(() => {
     if (!stripe) {
       return;
@@ -69,7 +74,44 @@ export default function CheckoutForm() {
       },
     });
 
+    // if(error.type === "validation_error"){
+    //   console.log("im in valid error");
+    //   setValidationError(true);
+    // }
 
+    if(error){
+      console.log("error in cart submit",error.type,validationError)
+      try {
+        async function completeOrder() {
+          if(error.type === "validation_error"){
+            //if a field is not filled,then stop.
+            return;
+          }else{
+            const result = await patchOrder(cart.id, {
+              id: cart.id,
+              userId: user.id,
+              isCart: false,
+              isComplete: true,
+              date: cart.order_date,
+            });
+            setCart(result.order);
+    
+            localStorage.removeItem("cart");
+    
+            // setClick(!click);
+            
+
+          }
+          
+        }
+        completeOrder();
+        
+      } catch (error) {
+        console.error(error);
+      }
+    }
+
+    console.log("this is inside stripe handlesubmit:",error)
     // This point will only be reached if there is an immediate error when
     // confirming the payment. Otherwise, your customer will be redirected to
     // your `return_url`. For some payment methods like iDEAL, your customer will
@@ -81,16 +123,27 @@ export default function CheckoutForm() {
       console.log(error)
       setMessage("An unexpected error occurred.");
     }
-
+   
     setIsLoading(false);
   };
 
   const paymentElementOptions = {
     layout: "tabs"
   }
+  // async function handleSubmitCart(e) {
+  //   e.preventDefault();
+  //   if (!stripe || !elements) {
+  //     // Stripe.js hasn't yet loaded.
+  //     // Make sure to disable form submission until Stripe.js has loaded.
+  //     return;
+  //   }
+
+   
+  // }
 
   return (
     <form id="payment-form stripe " onSubmit={handleSubmit}>
+      <h3>Payment</h3>
       <LinkAuthenticationElement
         id="link-authentication-element"
         onChange={(e) => setEmail(e.target.value)}
