+++ client/src/components/Shop/ItemRow.jsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from "react";
 import AddToCart from "./AddToCart";
 
 export default function ItemRow({ item }) {
-  console.log("item>itemRow>:",item)
+  
   return (
     <div>
       <img src={item.imagereel[0].image} alt="imageNotFound" />
