+++ client/src/components/Shop/Items.jsx
@@ -9,7 +9,7 @@ export default function Items({ items }) {
       <h2>Shop</h2>
       <div className="items-container">
         {items.map((item) => {
-          console.log("Items>itemmap:",item)
+          
           return (
             <div key={item.id} className="item-card">
               <Link key={item.id} to={`/shop/items/${item.id}`}>
